feat: migrate api to support cloud functions
