+++ lib/utils/constants.dart
@@ -17,9 +17,14 @@ const String masterDatabaseId = "64a521785f5be62b796f";
 const String roomsCollectionId = "64a5217e695bf2c4ec9c";
 const String participantsCollectionId = "64a63e508145d1084abf";
 
-//Pair chat database constants
+// Pair chat database constants
 const String pairRequestCollectionId = "64d980211f1395263ebe";
 const String activePairsCollectionId = "64d980cd65ff2e08ab97";
+
+// Room related cloud function constants
+const String createRoomServiceId = "651e2670b1e4a26e3cf1";
+const String joinRoomServiceId = "651e3d8fa35c690ed957";
+const String deleteRoomServiceId = "651e348775f28d84e11e";
 
 const String sendOtpFunctionID = "6513e9d40b57c6ec156f";
 const String verifyOtpFunctionID = "651303df122abc151bf3";
@@ -28,9 +33,6 @@ const String updateEmailFunctionID = "64b27d2e813dd152f0edz";
 
 const String emailVerificationDatabaseID = "64a7bfd6b09121548bfe";
 const String verificationCollectionID = "64a7c0100eabfe8d3844";
-
-// Resonate Backend Constants
-const String resonateApiUrl = "http://10.12.52.227:3000";
 
 // Github Constants
 const String githubRepoUrl = "https://github.com/AOSSIE-Org/Resonate";
