+++ pubspec.yaml
@@ -51,7 +51,6 @@ dependencies:
   animated_splash_screen: ^1.3.0
   flutter_native_splash: ^2.2.19
   page_transition: ^2.0.9
-  chopper: ^6.1.4
   http: 
   flutter_otp_text_field: ^1.1.1
   random_string: ^2.3.1
@@ -85,7 +84,6 @@ dev_dependencies:
   freezed_annotation: ^2.2.0
   build_runner: ^2.3.3
   json_serializable: ^6.6.2
-  chopper_generator: ^6.0.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
